refactor: import 경로 수정 및 생성자 간소화 (#1)
- import 경로에 /pages/를 추가하여 변경된 경로를 반영함.
- 생성자를 {Key? key}) : super(key: key);에서 ({super.key}); 형태로 간소화함.
