+++ fe/lib/adjustment/adjustment_page.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import '../home/home_page.dart';
-import '../user_page.dart';
+import '../pages/user_page.dart';
 
 class AdjustmentPage extends StatefulWidget {
-  const AdjustmentPage({Key? key}) : super(key: key);
+  const AdjustmentPage({super.key});
 
   @override
   _AdjustmentPageState createState() => _AdjustmentPageState();
